Refactored Config class to use Singleton pattern
Singleton pattern was applied to the Config class to ensure that only one instance of the class can be created during the runtime of the program. This will prevent the creation of multiple instances of the same class, which can lead to inconsistent behavior and waste of resources.

Also, the font initialization was moved to the __new__ method, which is called only once when the instance of the class is created.

Additionally, the pygame import was removed from the files that used the Config class and replaced with a reference to the Config class. This will allow for better separation of concerns and improved maintainability of the code.