Refactored classes in monsters.py and file structure
Refactored the following classes: Creature, Monster, Mage, Bat, and Boss, which improved code readability and efficiency in the main.py file.
The monsters.py file was removed and divided into three files: creature.py, monster.py, and enemy.py.
The Creature class is now an abstract class from which the Player and Monster classes inherit.
The Monster class inherits from the same classes as the Player class and use composition with Bullet class and it is the superclass for the Bat, Mage, and Boss classes.
The abstract classes creating the interface were divided into the drawable.py, collidable.py, and updatable.py files and combined into a single file named interfaces.py.
A constant was added in the config.py file for damage dealt, and the MONSTER_IMAGE constant was changed to BAT_IMAGE.
All changes have positively affected code readability and efficiency, and the game is now functioning correctly.