Implemented MenuCommand abstract class and apply SOLID principles to Button classes
Created an abstract class called MenuCommand to implement the Command pattern and achieve SOLID principles for buttons, each subclass of MenuCommand has a single responsibility for executing its action.
Implemented the Factory pattern for creating instances of the buttons.
Added comments to button.py, menu_commands.py, and menu.py for easier understanding of the code.
Added a method handle_button_click in menu.py to simplify the button actions.